pyproject.toml
[tool.pysen]
version = "0.10"

[tool.pysen.lint]
enable_black = true
enable_flake8 = true
enable_isort = true
enable_mypy = true
mypy_preset = "strict"
line_length = 88
py_version = "py310"

  [[tool.pysen.lint.mypy_targets]]
    paths = ["."]


[tool.poetry]
name = "harmonica-heat-alarm-client"
version = "0.1.0"
description = ""
authors = ["aoirint <aoirint@gmail.com>"]
readme = "README.md"
packages = [{include = "harmonica_heat_alarm_client"}]

[tool.poetry.dependencies]
python = "~3.11"
pyserial = "^3.5"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
pysen = "^0.10.5"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
flake8-bugbear = "^23.7.10"
mypy = "^1.4.1"
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
